+++ rsp_vision/console_application/app.py
@@ -13,8 +13,8 @@ from rsp_vision.analysis.spatial_freq_temporal_freq import (
 from rsp_vision.load.load_data import load_data, read_config_file
 from rsp_vision.objects.enums import PhotonType
 from rsp_vision.objects.photon_data import PhotonData
-from rsp_vision.SWC_blueprint_interactions.objects import SWC_Blueprint_Spec
-from rsp_vision.SWC_blueprint_interactions.save_data import save_data
+from rsp_vision.objects.SWC_Blueprint import SWC_Blueprint_Spec
+from rsp_vision.save.save_data import save_data
 
 
 CONFIG_PATH = config("CONFIG_PATH")
