+++ rsp_vision/objects/SWC_Blueprint.py
